airflow/requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-connector-python[pandas]
snowflake-sqlalchemy
SQLAlchemy
python-dotenv
matplotlib